ai-shopper/requirements.txt
requests
beautifulsoup4
pydantic
python-dotenv
openai
pytest
streamlit
